backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
mysqlclient>=2.2,<3.0
gunicorn>=23.0,<24.0
